+++ viewbind/UIViewBind/UIViewBindInterface.cs
@@ -1,8 +1,0 @@
-using System;
-using Msg;
-
-public interface UIViewBindAttr
-{
-    public void BindAttr(AttrType type, object obj, Action<AttrDescInfo> handler);
-    public void UnBindAttr(AttrType type, object obj);
-}